Add initial implementation of List
Context:
--------

The list is immutable, and is backed by a relaxed radix balanced
tree (RRB-Tree). The list allows elements to be added and
removed from both ends, as well as random-access reads and writes.
This data structure is sometimes called an immutable vector and is
present in languages like Scala and Clojure.

This implementation allows to get a mutable (linear) version from
an immutable list, and vice versa. For example this is useful in
the cases where you want to apply a series of mutations locally
before returning from a function the immutable version after the
mutations.

Changelog:
----------

- Add Size method.
- Add Nth method.
- Add AddLast method.
- Add Clone method.